use a BytesStorage for meta and a FileStorage for data, adjust tests
meta is usually a dict and serialized to (a few) bytes by the backend,
so the appropriate storage for that is a ByteStorage (accepts and return bytes).

data is better handled as a file, as it may get very big and we do not want
to keep it completely in memory. thus, the appropriate storage for that is a
FileStorage (accepts and returns open files).