close files at the right place, add docs about that, add rewind calls
remove unneeded close method / context manager support for TrackingFileWrapper.
reason: the caller who opens the real file is also responsible for closing it.

in general, if a caller opens a file, gives it to some function, then the caller
also needs to make sure it is closed later (NOT the called function).

if some function opens a file and returns it, then the caller is also responsible
for closing the open file it got.