indexing: simplify index update method
due to the way whoosh's update_document works, we can do this much simpler.

we just need to know which latest revids we have in the backend and
which revids we have in the index.
then we just update the set (backend_latest_revids - index_revids).

it doesn't matter whether the backend revid is from a new itemid (whoosh add)
or from a previously existing itemid (whoosh update will find the correct
document to update because itemid is a unique field). whoosh's update_document
can deal with both cases.

further, if we use delete_by_term, it does no harm if we give some revids that
won't be found in the index.

remove the commented time.sleep(1) call (i thought it was needed for AsyncWriter,
but this is not the case as whoosh will always immediately get a lock on the
index as there is nothing writing to the index in parallel).