simplify create/clear_revision: it is all just store_revision with the meta and data you want
for overwriting existing revisions, we need to be careful.
the code currently requires overwrite=True for this.

fixed acl checking:

write = be able to add new revisions to existing items (items with >= 1 revision)
create = be able to create new items (create the 1st revision of an item)
overwrite = be able to overwrite existing revisions in-place