some indexing middleware methods now yield Revisions (not revids - use rev.revid if you need that)
I tried to optimize Revision and Meta class by reusing information we already have:
e.g. if we already did a index lookup before creating a related revision, we can
give the whoosh doc to Revision.__init__ so it does not access the index again.

It is also given to Meta.__init__ there, so it does not need to load metadata
from the store if you are just requesting some meta key/value that is also in
the whoosh document.

If some non-stored meta key or data is accessed, it gets loaded from store.

Added some __cmp__ methods to make sorting and comparing of Revisions and Meta
possible.