we don't need an extract backend for Trash/*, just complicates stuff and is rather pointless
copied from etherpad:

i just thought about renaming items, router, trash, etc.

the first idea was that a rename (saving a latest revision with different name
than revs before) could potentially move the whole item (all revs) to another backend.

e.g. if one renames Foo to Trash/Foo and Trash is routed to another backend.

but then i noticed that we do not need to do that. we can just store only that
(renamed) latest rev to the Trash backend and keep everything else where it is,
which is much simpler.

of course that is not quite the idea of a separate trashbin backend, one would
rather expect everything related to same item to free up space in content
backend and allocate space in trash backend when an item is trashed. so we can
as well give up having that separate Trash backend and just use the Trash/*
part of the namespace for the trash.

freeing up space in the content backend would then happen when the "trashbin"
is emptied, by some admin action "destroy all trash" or "destroy some specific
trash item", then all itemids relating to whatever has a trashed latest-rev
would be destroyed.